+++ packages/core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.0
 
+- Feature: Add zoomLevel as plugin-agnostic map information to store.
 - Feature: Add core state variable for map's center position.
 
 ## 1.0.0